If statement for rm -rf FS

+++ unmake.sh
@@ -5,8 +5,11 @@ set -x
 
 docker-compose down -v
 docker swarm leave --force
-# a hack to do rm -rf on the host, regardless of who the user is
-docker run --rm -v $PWD:/hostFS -w /hostFS --entrypoint /bin/rm fnndsc/ubuntu-python3 -rf FS
+
+if [ -e "FS" ]; then
+  # a hack to do rm -rf on the host, regardless of who the user is
+  docker run --rm -v $PWD:/hostFS -w /hostFS --entrypoint /bin/rm fnndsc/ubuntu-python3 -rf FS
+fi
 
 { set +x; } 2> /dev/null
 
@@ -14,3 +17,4 @@ docker run --rm -v $PWD:/hostFS -w /hostFS --entrypoint /bin/rm fnndsc/ubuntu-py
 test "$(docker container inspect chris 2>&1 > /dev/null)" \
   '=' 'Error: No such container: chris' \
   && ! test -e FS
+
